# Click-Reserva/lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClickReserva API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message?: string;
}

export interface SuccessResponse {
  success: boolean;
  message?: string;
}

export interface LoginBody {
  email: string;
  password: string;
}

export type UserRole = (typeof UserRole)[keyof typeof UserRole];

export const UserRole = {
  professor: "professor",
  coordinator: "coordinator",
} as const;

export interface User {
  id: number;
  name: string;
  email: string;
  role: UserRole;
  blocked: boolean;
  totalAbsences: number;
}

export interface LoginResponse {
  user: User;
  token?: string;
}

export interface Room {
  id: number;
  number: string;
  name: string;
  capacity: number;
  computers: number;
  isActive: boolean;
  description?: string | null;
}

export interface CreateRoomBody {
  number: string;
  name: string;
  capacity: number;
  computers: number;
  isActive?: boolean;
  description?: string | null;
}

export type ReservationStatus =
  (typeof ReservationStatus)[keyof typeof ReservationStatus];

export const ReservationStatus = {
  pending: "pending",
  confirmed: "confirmed",
  realized: "realized",
  cancelled: "cancelled",
  no_show: "no_show",
  rejected: "rejected",
} as const;

export interface Reservation {
  id: number;
  professorId: number;
  professorName: string;
  roomId: number;
  roomNumber: string;
  roomName: string;
  date: string;
  startTime: string;
  endTime: string;
  subject: string;
  classGroup: string;
  status: ReservationStatus;
  confirmedPresence: boolean;
  confirmedAt?: string | null;
  createdAt: string;
}

export interface CreateReservationBody {
  roomId: number;
  date: string;
  startTime: string;
  endTime: string;
  subject: string;
  classGroup: string;
  /** Optional - coordinator can create on behalf of a professor */
  professorId?: number;
}

export interface BlockedSlot {
  id: number;
  roomId?: number | null;
  roomNumber?: string | null;
  roomName?: string | null;
  date?: string | null;
  startTime: string;
  endTime: string;
  reason: string;
  createdAt: string;
}

export interface CreateBlockedSlotBody {
  roomId?: number | null;
  date?: string | null;
  startTime: string;
  endTime: string;
  reason: string;
}

export type ProfessorRole = (typeof ProfessorRole)[keyof typeof ProfessorRole];

export const ProfessorRole = {
  professor: "professor",
  coordinator: "coordinator",
} as const;

export interface Professor {
  id: number;
  name: string;
  email: string;
  role: ProfessorRole;
  blocked: boolean;
  totalAbsences: number;
  blockedAt?: string | null;
  blockReason?: string | null;
  isActive: boolean;
}

export type CreateProfessorBodyRole =
  (typeof CreateProfessorBodyRole)[keyof typeof CreateProfessorBodyRole];

export const CreateProfessorBodyRole = {
  professor: "professor",
  coordinator: "coordinator",
} as const;

export interface CreateProfessorBody {
  name: string;
  email: string;
  password: string;
  role?: CreateProfessorBodyRole;
}

export interface UpdateProfessorBody {
  name?: string;
  email?: string;
  isActive?: boolean;
}

export interface DashboardStats {
  totalProfessors: number;
  activeProfessors: number;
  totalRooms: number;
  activeRooms: number;
  todayReservations: number;
  weekReservations: number;
  pendingReservations: number;
  blockedProfessors: number;
  totalAbsences: number;
}

export interface ConflictResult {
  hasConflict: boolean;
  conflictingReservation?: Reservation | null;
}

export interface Absence {
  id: number;
  professorId: number;
  professorName: string;
  reservationId: number;
  absenceDate: string;
  recordedAt: string;
}

export interface SystemConfig {
  absenceLimitForBlock: number;
  toleranceMinutes: number;
}

export type GetReservationsParams = {
  date?: string;
  roomId?: number;
  professorId?: number;
  status?: string;
};

export type GetBlockedSlotsParams = {
  roomId?: number;
  date?: string;
};

export type CheckConflictsParams = {
  roomId: number;
  date: string;
  startTime: string;
  endTime: string;
  excludeId?: number;
};

export type GetAbsencesParams = {
  professorId?: number;
};