# Click-Reserva/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: ClickReserva API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with email and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginBody"
      responses:
        "200":
          description: Authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user info
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /blocked-slots:
    get:
      operationId: getBlockedSlots
      tags: [reservations]
      summary: List all blocked time slots
      parameters:
        - name: roomId
          in: query
          schema:
            type: integer
        - name: date
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of blocked slots
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/BlockedSlot"
    post:
      operationId: createBlockedSlot
      tags: [reservations]
      summary: Create a blocked time slot (coordinator only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBlockedSlotBody"
      responses:
        "201":
          description: Blocked slot created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedSlot"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /blocked-slots/{slotId}:
    delete:
      operationId: deleteBlockedSlot
      tags: [reservations]
      summary: Delete a blocked time slot (coordinator only)
      parameters:
        - name: slotId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Blocked slot deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "403":
          description: Forbidden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /dashboard/conflicts:
    get:
      operationId: checkConflicts
      tags: [dashboard]
      summary: Check for conflicts given time/room
      parameters:
        - name: roomId
          in: query
          required: true
          schema:
            type: integer
        - name: date
          in: query
          required: true
          schema:
            type: string
        - name: startTime
          in: query
          required: true
          schema:
            type: string
        - name: endTime
          in: query
          required: true
          schema:
            type: string
        - name: excludeId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Conflict check result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ConflictResult"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
    LoginBody:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
      required:
        - user
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [professor, coordinator]
        blocked:
          type: boolean
        totalAbsences:
          type: integer
      required:
        - id
        - name
        - email
        - role
        - blocked
        - totalAbsences
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
        professorId:
          type: integer
        professorName:
          type: string
        roomId:
          type: integer
        roomNumber:
          type: string
        roomName:
          type: string
        date:
          type: string
        startTime:
          type: string
        endTime:
          type: string
        subject:
          type: string
        classGroup:
          type: string
        status:
          type: string
          enum: [pending, confirmed, realized, cancelled, no_show, rejected]
        confirmedPresence:
          type: boolean
        confirmedAt:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - professorId
        - professorName
        - roomId
        - roomNumber
        - roomName
        - date
        - startTime
        - endTime
        - subject
        - classGroup
        - status
        - confirmedPresence
        - createdAt
...
    BlockedSlot:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
          nullable: true
        roomNumber:
          type: string
          nullable: true
        roomName:
          type: string
          nullable: true
        date:
          type: string
          nullable: true
        startTime:
          type: string
        endTime:
          type: string
        reason:
          type: string
        createdAt:
          type: string
      required:
        - id
        - startTime
        - endTime
        - reason
        - createdAt
    CreateBlockedSlotBody:
      type: object
      properties:
        roomId:
          type: integer
          nullable: true
        date:
          type: string
          nullable: true
        startTime:
          type: string
        endTime:
          type: string
        reason:
          type: string
      required:
        - startTime
        - endTime
        - reason
...
    ConflictResult:
      type: object
      properties:
        hasConflict:
          type: boolean
        conflictingReservation:
          $ref: "#/components/schemas/Reservation"
          nullable: true
      required:
        - hasConflict
...