# Click-Reserva/lib/api-zod/src/generated/types/reservation.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * ClickReserva API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ReservationStatus } from "./reservationStatus";

export interface Reservation {
  id: number;
  professorId: number;
  professorName: string;
  roomId: number;
  roomNumber: string;
  roomName: string;
  date: string;
  startTime: string;
  endTime: string;
  subject: string;
  classGroup: string;
  status: ReservationStatus;
  confirmedPresence: boolean;
  confirmedAt?: string | null;
  createdAt: string;
}